requirements.txt
reflex==0.6.3
hume==0.7.2
aiofiles==24.1.0
eval_type_backport==0.2.0
httpx==0.27.2
anyio==4.6.2.post1
idna==3.10
sniffio==1.3.1
certifi==2024.8.30
httpcore==1.0.6
h11==0.14.0
pydantic==2.9.2
annotated-types==0.7.0
pydantic_core==2.23.4
typing_extensions==4.12.2
websockets==12.0
openai==1.52.0
distro==1.9.0
jiter==0.6.1
tqdm==4.66.5
colorama==0.4.6
python-dotenv==1.0.1
reflex-audio-capture==0.0.4
reflex-pyplot==0.1.3
matplotlib==3.9.2
